refactor(engine): reuse insert block errors
