+++ crates/engine/primitives/Cargo.toml
@@ -13,6 +13,7 @@ workspace = true
 [dependencies]
 # reth
 reth-evm = { workspace = true, optional = true }
+reth-execution-errors.workspace = true
 reth-execution-types.workspace = true
 reth-payload-primitives.workspace = true
 reth-payload-builder-primitives = { workspace = true, optional = true }
@@ -44,6 +45,7 @@ trie-debug = []
 std = [
     "dep:reth-evm",
     "dep:reth-payload-builder-primitives",
+    "reth-execution-errors/std",
     "reth-execution-types/std",
     "reth-ethereum-primitives/std",
     "reth-primitives-traits/std",
