+++ crates/engine/primitives/src/error.rs
@@ -1,7 +1,7 @@
 use alloc::boxed::Box;
 use alloy_rpc_types_engine::ForkchoiceUpdateError;
-use reth_errors::ConsensusError;
-use reth_storage_api::errors::ProviderError;
+use reth_errors::{BlockExecutionError, BlockValidationError, ConsensusError, ProviderError};
+use reth_execution_errors::InternalBlockExecutionError;
 
 /// Represents all error cases when handling a new payload.
 ///
@@ -48,18 +48,73 @@ impl BeaconForkChoiceUpdateError {
     }
 }
 
-/// Error returned by
-/// [`PayloadValidator::validate_block_post_execution_with_hashed_state`](crate::PayloadValidator::validate_block_post_execution_with_hashed_state).
-///
-/// Distinguishes a block that violates a post-execution consensus rule from an internal failure to
-/// load the state needed to run the check, so the engine does not mark a block invalid over a
-/// transient provider error.
+/// All error variants possible when inserting or validating a block.
 #[derive(Debug, thiserror::Error)]
-pub enum PostExecutionValidationError {
-    /// The block violates a post-execution consensus rule.
+pub enum InsertBlockErrorKind {
+    /// Block violated consensus rules.
     #[error(transparent)]
     Consensus(#[from] ConsensusError),
-    /// Loading the state required to validate the block failed.
+    /// Block execution failed.
+    #[error(transparent)]
+    Execution(#[from] BlockExecutionError),
+    /// Provider error.
     #[error(transparent)]
     Provider(#[from] ProviderError),
+    /// Other errors.
+    #[error(transparent)]
+    Other(#[from] Box<dyn core::error::Error + Send + Sync + 'static>),
+}
+
+impl InsertBlockErrorKind {
+    /// Returns whether the error was caused by an invalid block.
+    pub const fn is_validation_error(&self) -> bool {
+        matches!(self, Self::Consensus(_) | Self::Execution(BlockExecutionError::Validation(_)))
+    }
+
+    /// Returns an [`InsertBlockValidationError`] if the error is caused by an invalid block.
+    ///
+    /// Returns an [`InsertBlockFatalError`] if the error is caused by an error that is not
+    /// validation related or is otherwise fatal.
+    ///
+    /// This is intended to be used to determine if we should respond `INVALID` as a response when
+    /// processing a new block.
+    pub fn ensure_validation_error(
+        self,
+    ) -> Result<InsertBlockValidationError, InsertBlockFatalError> {
+        match self {
+            Self::Consensus(err) => Ok(InsertBlockValidationError::Consensus(err)),
+            Self::Execution(err) => match err {
+                BlockExecutionError::Validation(err) => {
+                    Ok(InsertBlockValidationError::Validation(err))
+                }
+                BlockExecutionError::Internal(error) => {
+                    Err(InsertBlockFatalError::BlockExecutionError(error))
+                }
+            },
+            Self::Provider(err) => Err(InsertBlockFatalError::Provider(err)),
+            Self::Other(err) => Err(InternalBlockExecutionError::Other(err).into()),
+        }
+    }
+}
+
+/// Error variants that are not caused by invalid blocks.
+#[derive(Debug, thiserror::Error)]
+pub enum InsertBlockFatalError {
+    /// A provider error.
+    #[error(transparent)]
+    Provider(#[from] ProviderError),
+    /// An internal or fatal block execution error.
+    #[error(transparent)]
+    BlockExecutionError(#[from] InternalBlockExecutionError),
+}
+
+/// Error variants that are caused by invalid blocks.
+#[derive(Debug, thiserror::Error)]
+pub enum InsertBlockValidationError {
+    /// Block violated consensus rules.
+    #[error(transparent)]
+    Consensus(#[from] ConsensusError),
+    /// Validation error, transparently wrapping [`BlockValidationError`].
+    #[error(transparent)]
+    Validation(#[from] BlockValidationError),
 }
